remove neatplot, update viz.py, add matplotlibrc
